+++ client/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added SI event for simple search ([#90](https://github.com/project-lux/lux-frontend/issues/90)).
+- Added SI event for facet events ([#90](https://github.com/project-lux/lux-frontend/issues/90)).
 
 ### Changed